commands: convert command functions to use cli context
When executing a command function, pass it a cli context rather than (cli,
config) for arguments.

This fixes the situation where the config argument seemed like it should have
been an actual instance config, but rather was the sambacc.config module. That
was both confusing and forced every command function to repeat the same dance
to actually get an instance config.  Now CLI options and generic configuration
is handled via the cli context.


Signed-off-by: John Mulligan <jmulligan@redhat.com>